% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ALFA_example.R
\name{ALFA_example}
\alias{ALFA_example}
\title{Get Path to Example Image}
\usage{
ALFA_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If NULL, the example files will be listed.}
}
\description{
ALFA comes bundled with a number of sample files in its inst/extdata directory. This function make them easy to access.
}
\examples{
ALFA_example("raw/img1.jpg")
input_dir <- ALFA_example("raw")
}
