% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{allKMers}
\alias{allKMers}
\title{Extract all k-mers of the given string}
\usage{
allKMers(str, k)
}
\arguments{
\item{str}{The string to extract k-mers from}

\item{k}{Size of each k-mer}
}
\value{
A string vector containing all k-mers
}
\description{
Extract all k-mers of the given string
}
