% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{countVectors}
\alias{countVectors}
\title{Create a count vector of k-mers in the two strings}
\usage{
countVectors(str1, str2, k)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{k}{size of k-mer}
}
\value{
a list containing two count vectors
}
\description{
Create a count vector of k-mers in the two strings
}
\examples{
alfa::countVectors("ATGTGAC", "ATGACTT", 3)

}
