% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{euclideanDistance}
\alias{euclideanDistance}
\title{Calculate the Euclidean distance between two
strings}
\usage{
euclideanDistance(str1, str2, k)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{k}{size of k-mers}
}
\value{
the Euclidean distance between str1 and str2
}
\description{
Calculate the Euclidean distance between two
strings
}
\examples{
alfa::euclideanDistance("ATCATC", "ATTATC", 3)

}
\references{
Zielezinski, A., Vinga, S., Almeida, J. et al. Alignment-free sequence
comparison: benefits, applications, and tools. Genome Biol 18, 186 (2017).
}
