% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{fastEuclideanDistance}
\alias{fastEuclideanDistance}
\title{Calculate the Euclidean distance between two strings
using pre-computed values}
\usage{
fastEuclideanDistance(countVec1, countVec2)
}
\arguments{
\item{countVec1}{the precomputed count vector for str1}

\item{countVec2}{the precomputed count vector for str2}
}
\value{
the Euclidean distance between countVec1 and countVec2
}
\description{
Calculate the Euclidean distance between two strings
using pre-computed values
}
