% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{fastStdEuclideanDistance}
\alias{fastStdEuclideanDistance}
\title{Calculate the standardized Euclidean distance between two
strings normalized by dividing by the standard deviation}
\usage{
fastStdEuclideanDistance(
  countVec1,
  countVec2,
  str1Len,
  str2Len,
  words,
  overlapCap,
  k
)
}
\arguments{
\item{countVec1}{count vector for the first string to be tested, generated
by countVectors()}

\item{countVec2}{count vector for the second string to be tested, generated
by countVectors()}

\item{str1Len}{length of the first string}

\item{str2Len}{length of the second string}

\item{words}{word factors that is the union of all possible k-mers from str1
and str2}

\item{overlapCap}{overlapping capability of each k-mer}

\item{k}{size of k-mers}
}
\value{
the standardized Euclidean distance between str1 and str2
}
\description{
Calculate the standardized Euclidean distance between two
strings normalized by dividing by the standard deviation
}
