% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenyBuilder.R
\name{neighborJoiningTree}
\alias{neighborJoiningTree}
\title{Construct phylogenetic tree using the neighbor-joining algorithm.
This function currently uses the function \code{nj} of the package \code{ape} to
build the tree internally.}
\usage{
neighborJoiningTree(distMat, ...)
}
\arguments{
\item{distMat}{distance matrix for constructing the phylogenetic tree}

\item{...}{additional parameters for the plotting function \code{plot}}
}
\value{
a phylogenetic tree object
}
\description{
Construct phylogenetic tree using the neighbor-joining algorithm.
This function currently uses the function \code{nj} of the package \code{ape} to
build the tree internally.
}
\examples{
library("ape")
woodmouse <- as.list(woodmouse)
woodmouse <- lapply(woodmouse, ape::as.character.DNAbin)
woodmouse <- lapply(woodmouse, paste0, collapse = "")
woodmouseSet <- Biostrings::DNAStringSet(unlist(woodmouse))
distMat <- alfa::createDistanceMatrix(woodmouseSet, metric = "euclidean")
alfa::neighborJoiningTree(distMat)

}
\references{
Paradis E, Schliep K (2019). “ape 5.0: an environment for modern
phylogenetics and evolutionary analyses in R.” Bioinformatics, 35, 526-528.
}
