% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenyBuilder.R
\name{plotDistanceCorrelation}
\alias{plotDistanceCorrelation}
\title{Plot the pairwise distance in the distance matrix against
the pairwise distance on the given phylogenetic tree.}
\usage{
plotDistanceCorrelation(distMat, tree)
}
\arguments{
\item{distMat}{pairwise distance matrix}

\item{tree}{phylogenetic tree to compare against}
}
\description{
This can be used to compare the effectiveness of a given alignment-free method
when \code{tree} is a phylogenetic tree created using alternative methods
such as MSA.
}
\examples{
library("ape")
data("woodmouse")
woodmouse <- as.list(woodmouse)
woodmouse <- lapply(woodmouse, ape::as.character.DNAbin)
woodmouse <- lapply(woodmouse, paste0, collapse = "")
woodmouseSet <- Biostrings::DNAStringSet(unlist(woodmouse))
distMat <- alfa::createDistanceMatrix(woodmouseSet, metric = "euclidean")
njtree <- alfa::neighborJoiningTree(distMat)
plotDistanceCorrelation(distMat, njtree)

}
