% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMatrix.R
\name{plotPairwiseDist}
\alias{plotPairwiseDist}
\title{Create and then plot the distance matrix as a heatmap}
\usage{
plotPairwiseDist(dnaSet, metric = "euclidean", k = 10, ...)
}
\arguments{
\item{dnaSet}{an object of type DNAStringSet or AAStringSet}

\item{metric}{metric for computing the distance, for now only 'euclidean'
and 'standardizedEuclidean' are supported}

\item{k}{size of k-mers (10 by default for DNA strings)}

\item{...}{arguments to be passed to the plotting function 'heatmap'}
}
\description{
Create and then plot the distance matrix as a heatmap
}
\examples{
library("ape")
woodmouse <- as.list(woodmouse)
woodmouse <- lapply(woodmouse, ape::as.character.DNAbin)
woodmouse <- lapply(woodmouse, paste0, collapse = "")
woodmouseSet <- Biostrings::DNAStringSet(unlist(woodmouse))
alfa::plotPairwiseDist(woodmouseSet, metric = "euclidean")

}
