% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Prepare an Image of Leaves for the Assessment of Leaf Area}
\usage{
preprocess(
  source,
  output_dir = NULL,
  crop = 0,
  red_scale = 0,
  mask_pixels = 0,
  mask_offset_x = 0,
  mask_offset_y = 0,
  workers = NULL
)
}
\arguments{
\item{source}{The path to the image, or directory of images, which you want to prepare for assessment.}

\item{output_dir}{The directory where to save the processed images, so that they can be assessed using \code{assess}. Respects tilde expansion. This must not be a directory that already exists, so that existing files are not over-written.}

\item{crop}{Should the margins of the image be cropped? Cropping occurs before the other operations, meaning that they are performed on the cropped image. Value is one integer; number of pixels to be removed from (top, bottom, left, right) of the image. Default = 0.}

\item{red_scale}{How many pixels wide should the side of the scale should be? Default = 0.}

\item{mask_pixels}{How many pixels should each side of the masking window be? Default = 0.}

\item{mask_offset_x}{Offset for positioning the masking window in number of pixels from right to left of the image.}

\item{mask_offset_y}{Offset for positioning the masking window in number of pixels from bottom to top of the image.}

\item{workers}{By default, preprocess will use all but one core for processing a folder of images. Here, you can control how many cores are used. Ignored when preprocessing a single image.}
}
\value{
No value is returned. The side effect is that the processed image is saved in the directory specified in \code{output_dir}.
}
\description{
\code{preprocess} prepares images of leaves for assessment using \code{assess}. Not all images require preprocessing. It provides options to crop away colored margins from an image, to mask out an un-needed scale bar, or to insert a scale bar of a particular size. Either a single image or single directory of images can be processed at once. If the latter, processing can be performed in parallel.
}
\examples{
img <- ALFA_example("raw/img4.jpg")
input_dir <- ALFA_example("raw")
output_dir <- ALFA_example("prepared")

\dontrun{
preprocess(img, output_dir, crop = -300), mask_pixels = 500, mask_offset_x = 0, mask_offset_y = 20)

preprocess(input_dir, output_dir, crop = 300, mask_pixels = 500, mask_offset_x = 0, mask_offset_y = 2300)
}
}
