% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_version.R
\name{python_version}
\alias{python_version}
\title{Determine the version of Python available}
\usage{
python_version()
}
\description{
This function assesses which version of Python is available. It first searches for Python 3, then for Python 2. The user is urged to update their Python installation if it is out of date.
#'
}
\examples{
python_version()



}
