% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMatrix.R
\name{readFASTA}
\alias{readFASTA}
\title{Read FASTA file into a DNAStringSet object to be processed by other
functions.}
\usage{
readFASTA(filepath)
}
\arguments{
\item{filepath}{path to the FASTA file to be imported}
}
\value{
a DNAStringSet object
}
\description{
Read FASTA file into a DNAStringSet object to be processed by other
functions.
}
\examples{
download.file(paste0("https://raw.githubusercontent.com/gaojunxuan/",
                     "alfa/main/inst/extdata/usflu.fasta"), "usaflu.fasta")
alfa::readFASTA("usaflu.fasta")

}
