% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_count.R
\name{sort_by_count}
\alias{sort_by_count}
\title{Sort images by count}
\usage{
sort_by_count(path, count)
}
\arguments{
\item{path}{Name of folder in which to sort images.}

\item{count}{How many images should be placed into each subdirectory?}
}
\description{
Some times it's convenient to sort images into directories such that every directory has the same number of images. This convenience function does that.
}
\examples{
input_dir <- ALFA_example("raw")
sort_by_count(input_dir, 2)
}
