% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_resolution.R
\name{sort_by_resolution}
\alias{sort_by_resolution}
\title{Sort images by resolution}
\usage{
sort_by_resolution(path)
}
\arguments{
\item{path}{Name of folder in which to sort images.}
}
\description{
Some image-processing programs assume that ones images are sorted by resolution. This helper function looks up the resolutions of each file, and moves them into folders. Files for which no resolution data is available in exif are placed in a folder called 'resolution_NA'.
}
\examples{
input_dir <- area_example("raw1")
sort_by_resolution(input_dir)
}
