% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{standardizedEuclidean}
\alias{standardizedEuclidean}
\title{Calculate the Euclidean distance between two
strings normalized by dividing by the standard deviation}
\usage{
standardizedEuclidean(str1, str2, k)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{k}{size of k-mers}
}
\value{
the standardized Euclidean distance between str1 and str2
}
\description{
Calculate the Euclidean distance between two
strings normalized by dividing by the standard deviation
}
\examples{
alfa::standardizedEuclidean("ACAATC", "ACATAA", 3)

}
\references{
Wu TJ, Burke JP, Davison DB. A measure of DNA sequence dissimilarity
based on Mahalanobis distance between frequencies of words.
Biometrics. 1997 Dec;53(4):1431-9. PMID: 9423258.
}
