% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{fastStdEuclideanDistance}
\alias{fastStdEuclideanDistance}
\title{Calculate the standardized Euclidean distance between two
strings normalized by dividing by the standard deviation}
\usage{
fastStdEuclideanDistance(
  countVec1,
  countVec2,
  str1Len,
  str2Len,
  words,
  overlapCap,
  k
)
}
\arguments{
\item{k}{size of k-mers}

\item{str1}{the first string}

\item{str2}{the second string}
}
\value{
the standardized Euclidean distance between str1 and str2
}
\description{
Calculate the standardized Euclidean distance between two
strings normalized by dividing by the standard deviation
}
