% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsStrDist.R
\name{hammingDistance}
\alias{hammingDistance}
\title{Calculate the Hamming distance of the two strings
(i.e. mean(count vector of str1 != count vector of str2))}
\usage{
hammingDistance(str1, str2, k)
}
\arguments{
\item{str1}{the first string}

\item{str2}{the second string}

\item{k}{size of k-mers}
}
\value{
the Hamming distance between str1 and str2
}
\description{
Calculate the Hamming distance of the two strings
(i.e. mean(count vector of str1 != count vector of str2))
}
