% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMatrix.R
\name{readFASTA}
\alias{readFASTA}
\title{Read FASTA file into a DNAStringSet object to be processed by other
functions.}
\usage{
readFASTA(filepath)
}
\arguments{
\item{filepath}{path to the FASTA file to be imported}
}
\value{
a DNAStringSet object
}
\description{
Read FASTA file into a DNAStringSet object to be processed by other
functions.
}
\examples{
alfa::readFASTA("inst/extdata/usflu.fasta")

}
